+++ feature-userinfo/build.gradle.kts
@@ -1,0 +1,63 @@
+plugins {
+    id("com.android.application")
+    id("org.jetbrains.kotlin.android")
+}
+
+android {
+    namespace = "com.example.feature_userinfo"
+    compileSdk = 34
+
+    defaultConfig {
+        applicationId = "com.example.feature_userinfo"
+        minSdk = 24
+        targetSdk = 34
+        versionCode = 1
+        versionName = "1.0"
+
+        testInstrumentationRunner = "androidx.test.runner.AndroidJUnitRunner"
+    }
+
+    buildTypes {
+        release {
+            isMinifyEnabled = false
+            proguardFiles(
+                getDefaultProguardFile("proguard-android-optimize.txt"),
+                "proguard-rules.pro"
+            )
+        }
+    }
+    compileOptions {
+        sourceCompatibility = JavaVersion.VERSION_17
+        targetCompatibility = JavaVersion.VERSION_17
+    }
+    kotlinOptions {
+        jvmTarget = "17"
+    }
+    composeOptions{
+        kotlinCompilerExtensionVersion = "1.5.0"
+    }
+    buildFeatures{
+        compose = true
+    }
+}
+
+dependencies {
+
+    //material3
+    implementation("androidx.compose.material3:material3:1.1.0")
+
+    //preview
+    implementation("androidx.compose.ui:ui-tooling-preview")
+
+    implementation("androidx.compose.ui:ui:1.1.0")
+
+    //bottom navigation
+    implementation("androidx.navigation:navigation-compose:2.7.0")
+
+    implementation("androidx.core:core-ktx:1.9.0")
+    implementation("androidx.appcompat:appcompat:1.6.1")
+    testImplementation("junit:junit:4.13.2")
+    androidTestImplementation("androidx.test.ext:junit:1.1.5")
+    androidTestImplementation("androidx.test.espresso:espresso-core:3.5.1")
+    debugImplementation("androidx.compose.ui:ui-tooling:1.6.0")
+}
