+++ feature-userinfo/proguard-rules.pro
