+++ settings.gradle.kts
@@ -23,5 +23,4 @@ include(":sample-app")
 include(":core-repository")
 include(":core-network")
 include(":authentication")
-include(":feature-userinfo")
-include(":userInfo")
+include(":feature-userInfo")
